feat: Prevent discontinuities in timeseries
If a check is configured with an interval longer than 5 minutes, the
generated time series will have discontinuities in them.

This PR adds logic to prevent those by repeating the last value
of a series at the required interval to prevent discontinuities.

This only affects metrics that are present between two consecutive
executions of a check.

Signed-off-by: Adrian Serrano <adrisr83@gmail.com>